internal: move config, controller, status packages
Moves these three packages up one level, from under
internal/operator/ to directly under internal/. This
simplifies the package hierarchy.
